pyproject.toml
[tool.poetry]
name = "label-studio"
version = "0.1.0"
description = ""
authors = ["Kei Shimizu <suns.manage.k@gmail.com>"]
readme = "README.md"
packages = [{include = "label_studio"}]

[tool.poetry.dependencies]
python = "^3.8"
label-studio-sdk = "^0.0.29"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
